refactor: consolidate logging and formatTimestamp, move evaluation tests to data-model
