--- library/data-model/src/derivedFields/logging.ts
@@ -1,24 +1,0 @@
-/*
- * Copyright 2026 Macquarie University
- *
- * Licensed under the Apache License Version 2.0 (the, "License");
- * you may not use, this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing software
- * distributed under the License is distributed on an "AS IS" BASIS
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND either express or implied.
- * See, the License, for the specific language governing permissions and
- * limitations under the License.
- *
- * Filename: logging.ts
- * Description:
- *   Warn logging for the derived-fields module. Swap the implementation here
- *   if data-model adopts a shared logging utility.
- */
-
-export const logWarn = (...args: unknown[]): void => {
-  console.warn(...args);
-};
