+++ library/data-model/tests/resolveParentField.test.ts
@@ -1,6 +1,5 @@
 import type {DataEngine, UiSpecModel} from '@faims3/data-model';
-import {describe, expect, it} from 'vitest';
-import {formatFieldValue, resolveParentFieldValue} from '@faims3/data-model';
+import {formatFieldValue, resolveParentFieldValue} from '../src';
 
 const meta = {
   annotation: {include: false, label: 'annotation'},
