+++ library/data-model/tests/resolveParentValues.test.ts
@@ -1,6 +1,5 @@
 import type {DataEngine, UiSpecModel} from '@faims3/data-model';
-import {describe, expect, it} from 'vitest';
-import {resolveParentValues} from '@faims3/data-model';
+import {resolveParentValues} from '../src';
 
 const meta = {
   annotation: {include: false, label: 'annotation'},
